Put macro expand in seperate file
