+++ Makefile
@@ -29,6 +29,9 @@ top10-with-macro: install
 top100: install
 	$(SCAN_ALL) data/crate-lists/top100.csv top100 -n 64
 
+top100-with-macro: install
+	$(SCAN_ALL) data/crate-lists/top100.csv top100 -n 64 --expand-macro
+
 top1000: install
 	$(SCAN_ALL) data/crate-lists/top1000.csv top1000 -n 64
