+++ data/results/top10_metadata.csv
@@ -1,11 +1,11 @@
 crate, effects, macros, macro LoC, conditional blocks, conditional LoC, skipped calls, skipped call LoC, skipped fn pointers, skipped pointers LoC, skipped other, skipped other LoC, unsafe traits, unsafe trait LoC, unsafe impls, unsafe impl LoC, public fns, public fns with effects, public total effects, audited fns, audited LoC, total LoC
-rand, 29, 182, 1176, 41, 3120, 0, 0, 0, 0, 105, 105, 0, 0, 0, 0, 65, 33, 204, 69, 704, 13020
-syn, 263, 2358, 7060, 56, 413, 0, 0, 0, 0, 48, 103, 0, 0, 3, 3, 1399, 909, 17678, 1592, 16308, 59726
-rand_core, 2, 19, 19, 3, 195, 0, 0, 0, 0, 8, 59, 0, 0, 0, 0, 24, 0, 0, 1, 5, 1474
+rand, 29, 182, 1176, 41, 3120, 0, 0, 0, 0, 105, 105, 0, 0, 0, 0, 65, 34, 232, 73, 664, 13061
+syn, 263, 2358, 7060, 56, 413, 0, 0, 0, 0, 48, 103, 0, 0, 3, 3, 1399, 725, 7689, 1216, 13088, 63149
+rand_core, 2, 19, 19, 3, 195, 0, 0, 0, 0, 8, 59, 0, 0, 0, 0, 24, 0, 0, 1, 5, 1483
 libc, 6818, 834, 45418, 2, 9, 0, 0, 0, 0, 18, 18, 0, 0, 2, 2, 6812, 6722, 6791, 6729, 13264, 125761
 cfg-if, 0, 8, 107, 1, 87, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 226
-quote, 1, 75, 2475, 0, 0, 0, 0, 0, 0, 17, 17, 0, 0, 0, 0, 111, 1, 1, 1, 3, 4238
-proc-macro2, 7, 94, 275, 79, 724, 0, 0, 0, 0, 13, 13, 0, 0, 0, 0, 70, 1, 1, 19, 91, 5516
+quote, 1, 75, 2475, 0, 0, 0, 0, 0, 0, 17, 17, 0, 0, 0, 0, 111, 1, 1, 1, 3, 4241
+proc-macro2, 7, 94, 275, 79, 724, 0, 0, 0, 0, 13, 13, 0, 0, 0, 0, 70, 1, 1, 12, 72, 5636
 unicode-xid, 0, 4, 4, 5, 29, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 1687
-serde, 16, 1971, 4718, 12, 99, 0, 0, 0, 0, 401, 435, 0, 0, 0, 0, 111, 24, 119, 180, 1411, 26165
-autocfg, 45, 53, 92, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 37, 24, 541, 35, 308, 874
+serde, 16, 1971, 4718, 12, 99, 0, 0, 0, 0, 401, 435, 0, 0, 0, 0, 111, 24, 123, 176, 1407, 27301
+autocfg, 45, 53, 92, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 37, 23, 891, 31, 297, 1028
