+++ data/results/top10_patterns.csv
@@ -1,8 +1,8 @@
 crate, [ClosureCreation], [FFI Call], [FFI Declaration], [FnPtrCreation], [PtrDeref], [StaticExtVar], [StaticMutVar], [UnsafeCall], std::env, std::ffi, std::fs, std::io, std::os, std::path, std::process
-rand, 0, 0, 0, 1, 4, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0
+rand, 0, 0, 0, 1, 4, 0, 0, 24, 0, 0, 0, 0, 0, 0, 0
 syn, 0, 3, 0, 223, 8, 0, 0, 29, 0, 0, 0, 0, 0, 0, 0
 rand_core, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0
-libc, 0, 23, 6605, 0, 0, 1, 1, 168, 0, 0, 0, 0, 0, 0, 0
+libc, 0, 23, 6605, 0, 0, 1, 1, 188, 0, 0, 0, 0, 0, 0, 0
 cfg-if, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 quote, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 proc-macro2, 0, 0, 0, 0, 0, 0, 0, 7, 0, 0, 0, 0, 0, 0, 0
