+++ src/scanner.rs
@@ -5,6 +5,7 @@
 
 use crate::attr_parser::CfgPred;
 use crate::audit_file::EffectInfo;
+use crate::macro_expand::handle_macro_expansion;
 use crate::resolution::hacky_resolver::HackyResolver;
 use crate::resolution::name_resolution::Resolver;
 
@@ -22,12 +23,6 @@ use petgraph::visit::{Bfs, EdgeRef};
 use petgraph::Direction;
 use proc_macro2::{TokenStream, TokenTree};
 use quote::ToTokens;
-use ra_ap_hir::db::ExpandDatabase;
-use ra_ap_ide::RootDatabase;
-use ra_ap_ide_db::syntax_helpers::insert_whitespace_into_node::insert_ws_into;
-use ra_ap_syntax::ast::{HasName, MacroCall};
-use ra_ap_syntax::{AstNode, SyntaxKind, SyntaxNode};
-use ra_ap_vfs::FileId;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::fmt::Debug;
@@ -54,6 +49,7 @@ pub struct ScanResults {
 
     pub call_graph: DiGraph<CanonicalPath, SrcLoc>,
     pub node_idxs: HashMap<CanonicalPath, NodeIndex>,
+    // To store call information for macro, Putting all call info in the could make some effect unrecognizable
     pub macro_call_graph: DiGraph<CanonicalPath, SrcLoc>,
     pub macro_node_idxs: HashMap<CanonicalPath, NodeIndex>,
 
@@ -263,6 +259,10 @@ where
 
     pub fn add_sinks(&mut self, new_sinks: HashSet<IdentPath>) {
         self.sinks.extend(new_sinks);
+    }
+
+    pub fn set_current_macro_context(&mut self, macro_context: Option<String>) {
+        self.current_macro_context = macro_context;
     }
 
     /*
@@ -475,7 +475,7 @@ where
                     self.data.skipped_macros.add(m);
                 }
                 syn::TraitItem::Verbatim(v) => {
-                     self.syn_info("skipping Verbatim expression", v);
+                    self.syn_info("skipping Verbatim expression", v);
                 }
                 other => {
                     self.data.skipped_other.add(other);
@@ -1367,212 +1367,19 @@ pub fn scan_file(
     scanner.add_sinks(sinks.clone());
     scanner.scan_file(&syntax_tree);
     if expand_macro {
-        let macro_hacky_resolver = HackyResolver::new(crate_name, filepath);
-        let mut macro_scanner = Scanner::new(
+        handle_macro_expansion(
+            crate_name,
             filepath,
-            macro_hacky_resolver.unwrap(),
+            resolver,
             macro_scan_results,
+            sinks,
             enabled_cfg,
-        );
-        macro_scanner.add_sinks(sinks);
-
-        let current_file_id =
-            resolver.find_file_id(&filepath).context("cannot find current file id")?;
-        let syntax = resolver.db().parse_or_expand(current_file_id.into());
-        let file_resolver_expand = FileResolver::new(crate_name, resolver, filepath)?;
-        let mut expanded_files: Vec<(syn::File, String)> = Vec::new();
-        let mut expanded_items: Vec<(syn::Item, String)> = Vec::new();
-        let mut expanded_stmts: Vec<(syn::Stmt, String)> = Vec::new();
-        let mut expanded_exprs: Vec<(syn::Expr, String)> = Vec::new();
-        let mut expanded_blocks: Vec<(syn::Block, String)> = Vec::new();
-
-        let ignored_macros: HashSet<&str> = [
-            "println",
-            "eprintln",
-            "dbg",
-            "assert",
-            "assert_eq",
-            "assert_ne",
-            "debug_assert",
-            "debug_assert_eq",
-            "debug_assert_ne",
-            "todo",
-            "unimplemented",
-            "cfg",
-            "cfg_attr",
-            "compile_error",
-            "info",
-            "warn",
-            "error",
-            "trace",
-            "debug",
-            "json",
-            "Serialize",
-            "Deserialize",
-            "tracing",
-            "tracing::info",
-            "tracing::warn",
-            "tracing::debug",
-            "tracing::trace",
-            "arg",
-            "command",
-            "test",
-            "bench",
-            "vec",
-        ]
-        .iter()
-        .cloned()
-        .collect();
-
-        for macro_call in find_macro_calls(&syntax) {
-            let macro_name;
-            if let Some(path) = macro_call.path() {
-                macro_name = path.syntax().text().to_string();
-                if ignored_macros.contains(macro_name.as_str()) {
-                    debug!("Ignored macro call: {}", macro_name);
-                    continue;
-                }
-            } else {
-                debug!("Failed to resolve macro path.");
-                continue;
-            }
-            let canonical_path = match get_canonical_path_from_ast(
-                filepath,
-                macro_call.syntax(),
-                macro_name,
-            ) {
-                Some(path) => path,
-                None => continue,
-            };
-            if let Some(expanded_syntax_node) =
-                file_resolver_expand.expand_macro(&macro_call)
-            {
-                let expansion = format(
-                    resolver.db(),
-                    macro_call
-                        .syntax()
-                        .parent()
-                        .map(|it| it.kind())
-                        .unwrap_or(SyntaxKind::MACRO_ITEMS),
-                    current_file_id,
-                    expanded_syntax_node,
-                );
-                // match all possible parsed syntax tree
-                match try_parse_expansion(&expansion) {
-                    Ok(ParseResult::File(parsed_file)) => {
-                        expanded_files.push((parsed_file, canonical_path.clone()));
-                    }
-                    Ok(ParseResult::Item(parsed_item)) => {
-                        expanded_items.push((parsed_item, canonical_path.clone()));
-                    }
-                    Ok(ParseResult::Stmt(parsed_stmt)) => {
-                        expanded_stmts.push((parsed_stmt, canonical_path.clone()));
-                    }
-                    Ok(ParseResult::Expr(parsed_expr)) => {
-                        expanded_exprs.push((parsed_expr, canonical_path.clone()));
-                    }
-                    Ok(ParseResult::Block(parsed_block)) => {
-                        expanded_blocks.push((parsed_block, canonical_path.clone()));
-                    }
-                    Err(e) => {
-                        debug!("Failed to parse expansion: {}", e);
-                        debug!(
-                            "Original macro call: {}; parent kind: {:#?}",
-                            macro_call,
-                            macro_call
-                                .syntax()
-                                .parent()
-                                .map(|it| it.kind())
-                                .unwrap_or(SyntaxKind::MACRO_ITEMS)
-                        );
-                        debug!("Expanded code: {}", expansion);
-                    }
-                }
-            }
-        }
-        for i in 0..expanded_files.len() {
-            macro_scanner.current_macro_context = Some(expanded_files[i].1.clone());
-            macro_scanner.scan_file(&expanded_files[i].0);
-            macro_scanner.current_macro_context = None;
-        }
-
-        for i in 0..expanded_items.len() {
-            macro_scanner.current_macro_context = Some(expanded_items[i].1.clone());
-            macro_scanner.scan_item(&expanded_items[i].0);
-            macro_scanner.current_macro_context = None;
-        }
-
-        for i in 0..expanded_stmts.len() {
-            macro_scanner.current_macro_context = Some(expanded_stmts[i].1.clone());
-            macro_scanner.scan_fn_statement(&expanded_stmts[i].0);
-            macro_scanner.current_macro_context = None;
-        }
-
-        for i in 0..expanded_exprs.len() {
-            macro_scanner.current_macro_context = Some(expanded_exprs[i].1.clone());
-            macro_scanner.scan_expr(&expanded_exprs[i].0);
-            macro_scanner.current_macro_context = None;
-        }
+        )
+        .unwrap_or_else(|err| {
+            info!("Failed to scan file: {} ({})", filepath.to_string_lossy(), err);
+        });
     }
     Ok(())
-}
-
-/// Enum to represent the possible parsing results
-enum ParseResult {
-    File(syn::File),
-    Item(syn::Item),
-    Stmt(syn::Stmt),
-    Expr(syn::Expr),
-    Block(syn::Block),
-}
-
-/// Try to parse the given string into different `syn` types
-fn try_parse_expansion(expansion: &str) -> Result<ParseResult> {
-    let mut error: Vec<_> = Vec::new();
-
-    // Attempt parsing as a full file
-    if let Ok(parsed_file) = syn::parse_file(expansion) {
-        return Ok(ParseResult::File(parsed_file));
-    }
-    error.push(syn::parse_file(expansion).err());
-
-    // Attempt parsing as a single item
-    if let Ok(parsed_item) = syn::parse_str::<syn::Item>(expansion) {
-        return Ok(ParseResult::Item(parsed_item));
-    }
-    error.push(syn::parse_str::<syn::Item>(expansion).err());
-
-    // Attempt parsing as a statement
-    if let Ok(parsed_stmt) = syn::parse_str::<syn::Stmt>(expansion) {
-        return Ok(ParseResult::Stmt(parsed_stmt));
-    }
-    error.push(syn::parse_str::<syn::Stmt>(expansion).err());
-
-    // Attempt parsing as an expression
-    if let Ok(parsed_expr) = syn::parse_str::<syn::Expr>(expansion) {
-        return Ok(ParseResult::Expr(parsed_expr));
-    }
-    error.push(syn::parse_str::<syn::Expr>(expansion).err());
-
-    // Attempt parsing as a Block
-    if let Ok(parsed_block) = syn::parse_str::<syn::Block>(expansion) {
-        return Ok(ParseResult::Block(parsed_block));
-    }
-    error.push(syn::parse_str::<syn::Block>(expansion).err());
-
-    let wrapped_stmts = format!("{{{}}}", expansion);
-    if let Ok(parsed_wrapped_stmts) = syn::parse_str::<syn::Stmt>(&wrapped_stmts) {
-        return Ok(ParseResult::Stmt(parsed_wrapped_stmts));
-    }
-    error.push(syn::parse_str::<syn::Stmt>(expansion).err());
-
-    let wrapped_file = format!("fn dummy() {{\n{}\n}}", expansion);
-    if let Ok(parsed_wrapped_file) = syn::parse_file(&wrapped_file) {
-        return Ok(ParseResult::File(parsed_wrapped_file));
-    }
-    error.push(syn::parse_file(expansion).err());
-    // If none of the parsers worked, return the raw input for debugging
-    Err(anyhow!("Failed to parse expansion: {:#?}", error))
 }
 
 /// Try to run scan_file, reporting any errors back to the user
@@ -1676,13 +1483,6 @@ pub fn scan_crate_with_sinks(
     Ok(scan_results)
 }
 
-fn find_macro_calls(syntax: &SyntaxNode) -> Vec<MacroCall> {
-    syntax
-        .descendants()
-        .filter_map(<ra_ap_syntax::ast::MacroCall as AstNode>::cast)
-        .collect()
-}
-
 /// Scan the supplied crate
 pub fn scan_crate(
     crate_path: &FilePath,
@@ -1743,130 +1543,4 @@ fn check_fn_for_effects(scan_results: &ScanResults, fn_: &CanonicalPath) -> bool
     }
 
     false
-}
-
-fn format(
-    db: &RootDatabase,
-    kind: SyntaxKind,
-    file_id: FileId,
-    expanded: SyntaxNode,
-) -> String {
-    let expansion = insert_ws_into(expanded).to_string();
-
-    _format(db, kind, file_id, &expansion).unwrap_or(expansion)
-}
-
-fn _format(
-    db: &RootDatabase,
-    kind: SyntaxKind,
-    file_id: FileId,
-    expansion: &str,
-) -> Option<String> {
-    use ra_ap_ide_db::base_db::{FileLoader, SourceDatabase};
-    // hack until we get hygiene working (same character amount to preserve formatting as much as possible)
-    const DOLLAR_CRATE_REPLACE: &str = "__r_a_";
-    let expansion = expansion.replace("$crate", DOLLAR_CRATE_REPLACE);
-    let (prefix, suffix) = match kind {
-        SyntaxKind::MACRO_PAT => ("fn __(", ": u32);"),
-        SyntaxKind::MACRO_EXPR | SyntaxKind::MACRO_STMTS => ("fn __() {", "}"),
-        SyntaxKind::MACRO_TYPE => ("type __ =", ";"),
-        _ => ("", ""),
-    };
-    let expansion = format!("{prefix}{expansion}{suffix}");
-
-    let &crate_id = db.relevant_crates(file_id).iter().next()?;
-    let edition = db.crate_graph()[crate_id].edition;
-
-    let mut cmd = std::process::Command::new(ra_ap_toolchain::rustfmt());
-    cmd.arg("--edition");
-    cmd.arg(edition.to_string());
-
-    let mut rustfmt = cmd
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-        .ok()?;
-
-    std::io::Write::write_all(&mut rustfmt.stdin.as_mut()?, expansion.as_bytes()).ok()?;
-
-    let output = rustfmt.wait_with_output().ok()?;
-    let captured_stdout = String::from_utf8(output.stdout).ok()?;
-
-    if output.status.success() && !captured_stdout.trim().is_empty() {
-        // let output = captured_stdout.replace(DOLLAR_CRATE_REPLACE, "$crate");
-        let output = captured_stdout.trim().strip_prefix(prefix)?;
-        let output = match kind {
-            SyntaxKind::MACRO_PAT => output
-                .strip_suffix(suffix)
-                .or_else(|| output.strip_suffix(": u32,\n);"))?,
-            _ => output.strip_suffix(suffix)?,
-        };
-        let trim_indent = ra_ap_stdx::trim_indent(output);
-        // tracing::debug!("expand_macro: formatting succeeded");
-        Some(trim_indent)
-    } else {
-        None
-    }
-}
-
-/// Get the canonical path of a function, method, or module containing the macro call.
-pub fn get_canonical_path_from_ast(
-    filepath: &FilePath,
-    macro_call: &SyntaxNode,
-    macro_name: String,
-) -> Option<String> {
-    let mut current_node = macro_call.clone();
-    let mut path_components = Vec::new();
-    let mut file_components = Vec::new();
-
-    while let Some(parent) = current_node.parent() {
-        current_node = parent;
-
-        // Case 1: Macro inside a function
-        if let Some(func) = ra_ap_syntax::ast::Fn::cast(current_node.clone()) {
-            if let Some(name) = func.name() {
-                path_components.push(name.text().to_string());
-                break; // Stop at the function level
-            }
-        }
-
-        // Case 2: Macro inside an `impl` block (method)
-        if let Some(impl_block) = ra_ap_syntax::ast::Impl::cast(current_node.clone()) {
-            if let Some(ty) = impl_block.self_ty() {
-                path_components.push(ty.syntax().text().to_string());
-            }
-        }
-
-        // Case 3: Macro at module level
-        if let Some(module) = ra_ap_syntax::ast::Module::cast(current_node.clone()) {
-            if let Some(name) = module.name() {
-                path_components.push(name.text().to_string());
-            }
-        }
-    }
-
-    let filepath = filepath.to_str();
-    if let Some(path) = filepath {
-        let components: Vec<&str> = path.split('/').collect();
-        if let Some(src_index) = components.iter().position(|&c| c == "src") {
-            if src_index > 0 {
-                // Insert the component before "src" at the head
-                file_components.insert(0, components[src_index - 1].to_string());
-                // Push all components after src to the tail
-                for after_src in &components[src_index + 1..] {
-                    if let Some(stripped) = after_src.strip_suffix(".rs") {
-                        let without_extension = stripped;
-                        file_components.push(without_extension.to_string());
-                    } else {
-                        file_components.push(after_src.to_string());
-                    }
-                }
-            }
-        }
-    }
-    path_components.reverse();
-    file_components.append(&mut path_components);
-    file_components.push(macro_name);
-    Some(file_components.join("::"))
 }
